+++ common/models/Income.php
@@ -74,7 +74,7 @@ class Income extends \soft\db\ActiveRecord
             if ($this->isNewRecord) {
                 $dollar_course = DollarCourse::find()->one();
                 $this->cost_type = self::COST_TYPE_USD;
-                $this->total = -1 * intval($this->cost * $this->weight);
+                $this->total = -1 * floatval($this->cost * $this->weight);
                 $this->unity_type_id = $this->productType->sizeType->id;
             }
             return true;
